--- src/test/java/org/example/AppTest.java
@@ -1,12 +1,0 @@
-package org.example;
-
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-public class AppTest {
-
-    @Test
-    public void testApp() {
-        Assert.assertTrue(true);
-    }
-}
